+++ Gmulator/Core.Gbc/GbcJoypad.cs
@@ -22,11 +22,11 @@ public class GbcJoypad
         gbc.CpuMap.Set(0x00, 0x00, 0xff00, 0xff00, Read00, Write00, RamType.Register, 1);
     }
 
-    private byte Read00(int a) => Status;
-    private void Write00(int a, byte v) => Status = v;
+    private int Read00(int a) => Status;
+    private void Write00(int a, int v) => Status = v;
 
-    private byte _status;
-    public byte Status
+    private int _status;
+    public int Status
     {
         get => _status;
         set
