+++ Gmulator/Core.Nes/Mappers/Mapper030.cs
@@ -7,11 +7,11 @@ public class Mapper030 : BaseMapper
         Reset();
     }
 
-    public override byte ReadChr(int addr) => base.ReadChr(addr);
+    public override int ReadChr(int addr) => base.ReadChr(addr);
 
-    public override byte ReadPrg(int addr) => base.ReadPrg(0x4000 * Prg[addr >> 14 & 1] + addr % 0x4000);
+    public override int ReadPrg(int addr) => base.ReadPrg(0x4000 * Prg[addr >> 14 & 1] + addr % 0x4000);
 
-    public override byte ReadVram(int addr) => base.ReadVram(addr);
+    public override int ReadVram(int addr) => base.ReadVram(addr);
 
     public override void Reset()
     {
@@ -19,7 +19,7 @@ public class Mapper030 : BaseMapper
         Chr = [0];
     }
 
-    public override void Write(int addr, byte value)
+    public override void Write(int addr, int value)
     {
         if (addr >= 0xc000)
         {
@@ -28,5 +28,5 @@ public class Mapper030 : BaseMapper
         }
     }
 
-    public override void WritePrg(int addr, byte value) => base.WritePrg(addr, value);
+    public override void WritePrg(int addr, int value) => base.WritePrg(addr, value);
 }
